nb/intel/common/fixed_bars.h: Add casts to `uintptr_t`
64-bit builds need this, as the Kconfig values fit in a 32-bit integer.

Tested with BUILD_TIMELESS=1, Asrock B85M Pro4 remains identical.

Reviewed-on: https://review.coreboot.org/c/coreboot/+/50469
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>

+++ src/northbridge/intel/common/fixed_bars.h
@@ -9,16 +9,16 @@ _Static_assert(CONFIG_FIXED_EPBAR_MMIO_BASE  != 0,  "EPBAR base address is zero"
 
 #include <stdint.h>
 
-#define MCHBAR8(x)	(*((volatile u8  *)(CONFIG_FIXED_MCHBAR_MMIO_BASE + (x))))
-#define MCHBAR16(x)	(*((volatile u16 *)(CONFIG_FIXED_MCHBAR_MMIO_BASE + (x))))
-#define MCHBAR32(x)	(*((volatile u32 *)(CONFIG_FIXED_MCHBAR_MMIO_BASE + (x))))
+#define MCHBAR8(x)	(*((volatile u8  *)((uintptr_t)CONFIG_FIXED_MCHBAR_MMIO_BASE + (x))))
+#define MCHBAR16(x)	(*((volatile u16 *)((uintptr_t)CONFIG_FIXED_MCHBAR_MMIO_BASE + (x))))
+#define MCHBAR32(x)	(*((volatile u32 *)((uintptr_t)CONFIG_FIXED_MCHBAR_MMIO_BASE + (x))))
 
-#define DMIBAR8(x)	(*((volatile u8  *)(CONFIG_FIXED_DMIBAR_MMIO_BASE + (x))))
-#define DMIBAR16(x)	(*((volatile u16 *)(CONFIG_FIXED_DMIBAR_MMIO_BASE + (x))))
-#define DMIBAR32(x)	(*((volatile u32 *)(CONFIG_FIXED_DMIBAR_MMIO_BASE + (x))))
+#define DMIBAR8(x)	(*((volatile u8  *)((uintptr_t)CONFIG_FIXED_DMIBAR_MMIO_BASE + (x))))
+#define DMIBAR16(x)	(*((volatile u16 *)((uintptr_t)CONFIG_FIXED_DMIBAR_MMIO_BASE + (x))))
+#define DMIBAR32(x)	(*((volatile u32 *)((uintptr_t)CONFIG_FIXED_DMIBAR_MMIO_BASE + (x))))
 
-#define EPBAR8(x)	(*((volatile u8  *)(CONFIG_FIXED_EPBAR_MMIO_BASE + (x))))
-#define EPBAR16(x)	(*((volatile u16 *)(CONFIG_FIXED_EPBAR_MMIO_BASE + (x))))
-#define EPBAR32(x)	(*((volatile u32 *)(CONFIG_FIXED_EPBAR_MMIO_BASE + (x))))
+#define EPBAR8(x)	(*((volatile u8  *)((uintptr_t)CONFIG_FIXED_EPBAR_MMIO_BASE + (x))))
+#define EPBAR16(x)	(*((volatile u16 *)((uintptr_t)CONFIG_FIXED_EPBAR_MMIO_BASE + (x))))
+#define EPBAR32(x)	(*((volatile u32 *)((uintptr_t)CONFIG_FIXED_EPBAR_MMIO_BASE + (x))))
 
 #endif	/* ! NORTHBRIDGE_INTEL_COMMON_FIXED_BARS_H */
